Renamed TestJ2ClientItemMap for consistency reasons.
To J2ClientItemMapTest, like all other tests are named.
